Makes DarkModeButton Composable for reusability.
